Resolved issue with competing types.

+++ test/test_xfft.cpp
@@ -10,8 +10,8 @@ namespace xt
         size_t k = 2;
         size_t n = 8096;
         size_t A = 10;
-        auto x = xt::linspace<float>(0, static_cast<float>(n - 1), n);
-        xt::xarray<float> y = A * xt::sin(2 * xt::numeric_constants<float>::PI * x * k / n);
+        auto x = xt::linspace<double>(0, static_cast<double>(n - 1), n);
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
         auto res = xt::fft::fft(y) / (n / 2);
         REQUIRE(A == doctest::Approx(std::abs(res(k))).epsilon(.0001));
     }
@@ -21,17 +21,17 @@ namespace xt
         size_t k = 2;
         size_t n = 8;
         size_t A = 10;
-        auto x = xt::linspace<float> (0, static_cast<float>(n - 1), n);
-        xt::xarray<float> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
+        auto x = xt::linspace<double> (0, static_cast<double>(n - 1), n);
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
         auto res = xt::fft::ifft(y) / (n / 2);
         REQUIRE(A == doctest::Approx(std::abs(res(k))).epsilon(.0001));
     }
 
     TEST(xfft, convolve_power_2)
     {
-        xt::xarray<float> x = { 1.0, 1.0, 1.0, 5.0};
-        xt::xarray<float> y = { 5.0, 1.0, 1.0, 1.0};
-        xt::xarray<float> expected = { 12, 12, 12, 28 };
+        xt::xarray<double> x = { 1.0, 1.0, 1.0, 5.0};
+        xt::xarray<double> y = { 5.0, 1.0, 1.0, 1.0};
+        xt::xarray<double> expected = { 12, 12, 12, 28 };
 
         auto result = xt::fft::fftconvolve(x, y);
 
@@ -47,8 +47,8 @@ namespace xt
         size_t n = 15;
         size_t A = 1;
         size_t dim = 100;
-        auto x = xt::linspace<float>(0, n - 1, n) * xt::ones<float>({dim, n});
-        xt::xarray<float> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
+        auto x = xt::linspace<double>(0, n - 1, n) * xt::ones<double>({dim, n});
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
         y = xt::transpose(y);
         auto res = xt::fft::fft(y, 0) / (n / 2.0);
         REQUIRE(A == doctest::Approx(std::abs(res(k, 0))).epsilon(.0001));
@@ -61,8 +61,8 @@ namespace xt
         size_t n = 15;
         size_t A = 1;
         size_t dim = 2;
-        auto x = xt::linspace<float>(0, n - 1, n) * xt::ones<float>({ dim, n });
-        xt::xarray<float> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
+        auto x = xt::linspace<double>(0, n - 1, n) * xt::ones<double>({ dim, n });
+        xt::xarray<double> y = A * xt::sin(2 * xt::numeric_constants<double>::PI * x * k / n);
         auto res = xt::fft::fft(y) / (n / 2.0);
         REQUIRE(A == doctest::Approx(std::abs(res(0, k))).epsilon(.0001));
         REQUIRE(A == doctest::Approx(std::abs(res(1, k))).epsilon(.0001));
@@ -70,13 +70,13 @@ namespace xt
 
     TEST(xfft, convolve_n)
     {
-        xt::xarray<float> x = { 1.0, 1.0, 1.0, 5.0, 1.0 };
-        xt::xarray<float> y = { 5.0, 1.0, 1.0, 1.0, 1.0 };
+        xt::xarray<double> x = { 1.0, 1.0, 1.0, 5.0, 1.0 };
+        xt::xarray<double> y = { 5.0, 1.0, 1.0, 1.0, 1.0 };
         xt::xarray<size_t> expected = { 13, 13, 13, 29, 13 };
 
         auto result = xt::fft::fftconvolve(x, y);
 
-        xt::xarray<float> abs = xt::abs(result);
+        xt::xarray<double> abs = xt::abs(result);
         
         for(size_t i = 0; i < abs.size(); i++)
         {
